bindgen: cross-platform integrity (Option 6) — From<{u32,i32}> on core enums + constant overrides
Windows MSVC builds were failing with two distinct error classes
that don't reproduce on macOS clang:

1. 'i32 -> u32 mismatched type' on sites passing native::WGPU*
   enum values into core::*::from_raw(u32). Root cause:
   webgpu.h's WGPU_ENUM_ATTRIBUTE is an empty macro, so each C
   compiler picks the enum underlying type per its own defaults
   (macOS clang: c_uint=u32 for non-negative-only enums;
   Windows MSVC: c_int=i32 unconditionally). bindgen reflects
   libclang's choice verbatim.

2. 'cannot find value WGPU_WHOLE_MAP_SIZE' / similar. Root cause:
   #define X (SIZE_MAX) / (UINT64_MAX) cast macros are evaluated
   to -1 by bindgen's static evaluator and emitted as i32 = -1
   on macOS (numerically-OK after 'as usize' workarounds) and
   either dropped or differently typed on MSVC.

Fix is structural, not workaround-based: the platform-dep enum
type difference is absorbed inside core::* via From<u32> and
From<i32> impl pairs; the FFI layer (yawgpu/src/conv.rs) calls
'.into()' and the compiler routes to the right impl per
platform. yawgpu-core gains the impls but does NOT depend on
yawgpu's bindgen output (no native::* references leak into
core) - layering is preserved. Bit-preserving 'as' casts are
the only platform-dep code, localized to the From impl bodies.

Part A - yawgpu-core/src/lib.rs (+103 lines, test-bearing
public API additions per block-90):
- impl From<u32> / From<i32> for TextureFormat (and reverse).
- impl From<u32> / From<i32> for VertexFormat (and reverse).
- impl From<u32> / From<i32> for QueryType (and reverse) -
  match-based since QueryType is an enum-with-Unknown(u32).
- VertexFormat gained a 'pub fn raw(self) -> u32' getter
  that was previously missing.

Part B - yawgpu/src/conv.rs: rewrote map_texture_format /
map_texture_format_to_native / map_query_type /
map_query_type_to_native to one-liner '.into()' calls. Added
new map_vertex_format / map_vertex_format_to_native (previously
the conversion was inline at the call site). Routed all direct
core::*::from_raw(native::WGPU*) call sites through .into().

Part C - yawgpu/build.rs: added .clang_macro_fallback() (matches
wgpu-native) plus blocklist+raw_line overrides for the four
(SIZE_MAX)/(UINT64_MAX)-derived constants
(WGPU_WHOLE_MAP_SIZE/WHOLE_SIZE/LIMIT_U64_UNDEFINED/STRLEN),
pinning them to usize=usize::MAX / u64=u64::MAX so the bindings
are byte-for-byte identical across platforms.

Part D - call-site cleanup: dropped now-redundant 'as usize' /
'as u64' workaround casts at 12+ sites (caught by clippy
unnecessary_cast):
- yawgpu/src/conv.rs (1 site).
- yawgpu/src/lib.rs (2 sites in mapped_range fns).
- yawgpu/tests/{buffer_map,buffer_mapped_range,limits,
  device_lost}_validation.rs (9 sites).

Part E - block-90 unit-test coverage for the 6 new pub
'From' impls + new VertexFormat::raw getter:
- yawgpu-core/src/lib.rs tests gained assertions for each
  conversion direction (u32/i32 in/out, known + Unknown variant).
- specs/tracking/phase-10-coverage.md extended with rows for
  the new pub fns; phase-10-audit.md kept-pub distribution
  updated to reflect the +7 pub fns in yawgpu-core (TextureFormat
  raw was already pub; VertexFormat raw is new).

Part F - specs/reference/dependencies.md: documents the bindgen
workaround in the existing webgpu.h / bindgen 0.72 section.

Verification (macOS, 2026-05-21):
- Noop cargo test --workspace 499 -> 502 (+3 expected from new
  From-impl assertions threaded into existing tests +
  VertexFormat::raw coverage). 58 binaries unchanged.
- Workspace clippy + cargo clippy -p yawgpu --features metal/
  vulkan all clean.
- Regenerated bindings show the four constants correctly typed:
  WGPU_WHOLE_MAP_SIZE: usize = usize::MAX,
  WGPU_WHOLE_SIZE / WGPU_LIMIT_U64_UNDEFINED: u64 = u64::MAX,
  WGPU_STRLEN: usize = usize::MAX.
- Phase-7 Metal e2e regression sample (basic 3/3 + render 3/3)
  green.
- yawgpu-hal Vulkan --ignored sweep 22/22 green (with
  $VULKAN_SDK env).
- Phase-9 examples regression: Metal triangle exit 0.

Windows verification is pending on the user's MSVC host.
Structurally the fix removes both error classes:
- Enum type platform-difference is now absorbed inside the From
  impls (the compiler picks From<u32> on macOS, From<i32> on
  Windows).
- The four problem constants are pinned by raw_line so the
  Windows binding has them with the correct usize/u64 type.

This design is robust to future bindgen behavior changes
(e.g. switching to .newtype_enum or .rustified_enum): the core
API stays the same and only the From impls would need
adjustment.

+++ yawgpu/tests/buffer_map_validation.rs
@@ -257,7 +257,7 @@ fn valid_map_transitions_pending_then_mapped_success() {
             buffer,
             native::WGPUMapMode_Read,
             0,
-            native::WGPU_WHOLE_MAP_SIZE as usize,
+            native::WGPU_WHOLE_MAP_SIZE,
             native::WGPUCallbackMode_AllowProcessEvents,
             &mut state,
         );
